Refactor: Move service implementations to app/services
